Display a confirmation screen on delete

+++ plugins/tracker/include/Tracker/Tracker.class.php
@@ -576,7 +576,8 @@ class Tracker implements Tracker_Dispatchable_Interface {
                     $GLOBALS['Response']->redirect(TRACKER_BASE_URL.'/?tracker='. $this->getId());
                 }
                 break;
-            case 'admin-clean':
+
+           case 'admin-clean':
                 if ($this->userIsAdmin($current_user)) {
                     $this->displayAdminClean($layout);
                 } else {
@@ -584,14 +585,35 @@ class Tracker implements Tracker_Dispatchable_Interface {
                     $GLOBALS['Response']->redirect(TRACKER_BASE_URL.'/?tracker='. $this->getId());
                 }
                 break;
-            case 'admin-delete-artifact':
+
+           case 'admin-delete-artifact-confirm':
                 if ($this->userIsAdmin($current_user)) {
-                    $artifact = $this->getTrackerArtifactFactory()->getArtifactById($request->get('id'));
-                    if ($artifact && $artifact->getTrackerId() == $this->getId()) {
-                        $artifact->delete($current_user);
-                        $GLOBALS['Response']->addFeedback('info', $GLOBALS['Language']->getText('plugin_tracker_admin', 'clean_info_deleted', array($request->get('id'))));
+                    $artifact_id = $request->getValidated('id', 'uint', 0);
+                    $artifact    = $this->getTrackerArtifactFactory()->getArtifactById($artifact_id);
+                    if ($artifact) {
+                        $this->displayAdminConfirmDelete($layout, $artifact);
                     } else {
                         $GLOBALS['Response']->addFeedback('error', $GLOBALS['Language']->getText('plugin_tracker_admin', 'clean_error_noart', array($request->get('id'))));
+                        $GLOBALS['Response']->redirect(TRACKER_BASE_URL.'/?tracker='. $this->getId().'&func=admin-clean');
+                    }
+                } else {
+                    $GLOBALS['Response']->addFeedback('error', $GLOBALS['Language']->getText('plugin_tracker_admin', 'access_denied'));
+                    $GLOBALS['Response']->redirect(TRACKER_BASE_URL.'/?tracker='. $this->getId());
+                }
+                break;
+
+            case 'admin-delete-artifact':
+                if ($this->userIsAdmin($current_user)) {
+                    if ($request->exist('confirm')) {
+                        $artifact = $this->getTrackerArtifactFactory()->getArtifactById($request->get('id'));
+                        if ($artifact && $artifact->getTrackerId() == $this->getId()) {
+                            $artifact->delete($current_user);
+                            $GLOBALS['Response']->addFeedback('info', $GLOBALS['Language']->getText('plugin_tracker_admin', 'clean_info_deleted', array($request->get('id'))));
+                        } else {
+                            $GLOBALS['Response']->addFeedback('error', $GLOBALS['Language']->getText('plugin_tracker_admin', 'clean_error_noart', array($request->get('id'))));
+                        }
+                    } else {
+                        $GLOBALS['Response']->addFeedback('info', $GLOBALS['Language']->getText('plugin_tracker_admin', 'clean_cancel_deleted'));
                     }
                     $GLOBALS['Response']->redirect(TRACKER_BASE_URL.'/?tracker='. $this->getId().'&func=admin');
                 } else {
@@ -599,6 +621,7 @@ class Tracker implements Tracker_Dispatchable_Interface {
                     $GLOBALS['Response']->redirect(TRACKER_BASE_URL.'/?tracker='. $this->getId());
                 }
                 break;
+
             default:
                 $nothing_has_been_done = true;
                 EventManager::instance()->processEvent(
@@ -1763,14 +1786,30 @@ EOS;
     public function displayAdminClean(Tracker_IDisplayTrackerLayout $layout) {
         $this->displayAdminItemHeader($layout, 'clean');
         echo '<p>'.$GLOBALS['Language']->getText('plugin_tracker_admin', 'clean_info').'</p>';
-        echo '<form name="delete_artifact" method="post" action="'.TRACKER_BASE_URL.'/?tracker='. (int)$this->id.'&amp;func=admin-delete-artifact">';
+        echo '<form name="delete_artifact" method="post" action="'.TRACKER_BASE_URL.'/?tracker='. (int)$this->id.'&amp;func=admin-delete-artifact-confirm">';
         echo '<label>'.$GLOBALS['Language']->getText('plugin_tracker_admin', 'clean_id').' <input type="text" name="id" value=""></label>';
         echo '<br>';
         echo '<input type="submit" value="'.$GLOBALS['Language']->getText('global','btn_submit').'">';
         echo '</form>';
         $this->displayFooter($layout);
     }
-    
+
+    public function displayAdminConfirmDelete(Tracker_IDisplayTrackerLayout $layout, Tracker_Artifact $artifact) {
+        $this->displayAdminItemHeader($layout, 'clean');
+        echo '<div class="tracker_confirm_delete">';
+        echo '<form name="delete_artifact" method="post" action="'.TRACKER_BASE_URL.'/?tracker='. (int)$this->id.'&amp;func=admin-delete-artifact">';
+        echo $GLOBALS['Language']->getText('plugin_tracker_admin', 'clean_confirm_text', array($artifact->getXRefAndTitle()), CODENDI_PURIFIER_DISABLED);
+        echo '<br>';
+        echo '<input type="hidden" name="id" value="'.$artifact->getId().'" />';
+        echo '<div class="tracker_confirm_delete_buttons">';
+        echo '<input type="submit" tabindex="2" name="confirm" value="'. $GLOBALS['Language']->getText('plugin_tracker_admin', 'clean_confirm') .'" />';
+        echo '<input type="submit" tabindex="1" name="cancel" value="'. $GLOBALS['Language']->getText('plugin_tracker_admin', 'clean_cancel') .'" />';        
+        echo '</div>';
+        echo '</form>';
+        echo '</div>';
+        $this->displayFooter($layout);
+    }
+
     public function displayMasschangeForm(Tracker_IDisplayTrackerLayout $layout, $masschange_aids) {
         $breadcrumbs = array(
                 array(
